requirements.txt
build==0.7.0
packaging==21.3
pep517==0.12.0
pyparsing==3.0.9
python-xlib==0.31
six==1.16.0
tomli==2.0.1
